lavf/isom: Make auxiliary_offsets consistently uint64_t.
Fixes a compilation warning if size_t != uint64_t:
libavformat/mov.c: In function ‘mov_read_saio’:
libavformat/mov.c:6207:45: warning: assignment from incompatible pointer type [-Wincompatible-pointer-types]
         encryption_index->auxiliary_offsets = auxiliary_offsets;
                                             ^

+++ libavformat/isom.h
@@ -118,7 +118,7 @@ typedef struct MOVEncryptionIndex {
     uint8_t* auxiliary_info_sizes;
     size_t auxiliary_info_sample_count;
     uint8_t auxiliary_info_default_size;
-    size_t* auxiliary_offsets;  ///< Absolute seek position
+    uint64_t* auxiliary_offsets;  ///< Absolute seek position
     size_t auxiliary_offsets_count;
 } MOVEncryptionIndex;
